+++ src/main/java/com/webank/webase/sign/api/service/KeyStoreService.java
@@ -21,10 +21,11 @@ import com.webank.webase.sign.pojo.bo.KeyStoreInfo;
 import com.webank.webase.sign.util.AesUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.fisco.bcos.sdk.crypto.CryptoSuite;
-import org.fisco.bcos.sdk.crypto.keypair.CryptoKeyPair;
-import org.fisco.bcos.sdk.model.CryptoType;
+import org.fisco.bcos.sdk.v3.crypto.CryptoSuite;
+import org.fisco.bcos.sdk.v3.crypto.keypair.CryptoKeyPair;
+import org.fisco.bcos.sdk.v3.model.CryptoType;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
@@ -36,6 +37,12 @@ import org.springframework.stereotype.Service;
 public class KeyStoreService {
     @Autowired
     private AesUtils aesUtils;
+    @Autowired
+    @Qualifier(value = "sm")
+    private CryptoSuite smCryptoSuite;
+    @Autowired
+    @Qualifier(value = "ecdsa")
+    private CryptoSuite ecdsaCryptoSuite;
 
     /**
      * get KeyStoreInfo by privateKey.
@@ -87,17 +94,17 @@ public class KeyStoreService {
 
     public CryptoKeyPair getKeyPairByType(String privateKeyRaw, int encryptType) {
         if (encryptType == CryptoType.SM_TYPE) {
-            return new CryptoSuite(CryptoType.SM_TYPE).createKeyPair(privateKeyRaw);
+            return smCryptoSuite.getKeyPairFactory().createKeyPair(privateKeyRaw);
         } else {
-            return new CryptoSuite(CryptoType.ECDSA_TYPE).createKeyPair(privateKeyRaw);
+            return ecdsaCryptoSuite.getKeyPairFactory().createKeyPair(privateKeyRaw);
         }
     }
 
     public CryptoKeyPair getKeyPairRandom(int encryptType) {
         if (encryptType == CryptoType.SM_TYPE) {
-            return new CryptoSuite(CryptoType.SM_TYPE).getKeyPairFactory().generateKeyPair();
+            return smCryptoSuite.getKeyPairFactory().generateKeyPair();
         } else {
-            return new CryptoSuite(CryptoType.ECDSA_TYPE).getKeyPairFactory().generateKeyPair();
+            return ecdsaCryptoSuite.getKeyPairFactory().generateKeyPair();
         }
     }