+++ src/main/java/com/webank/webase/sign/api/service/SignService.java
@@ -21,20 +21,20 @@ import com.webank.webase.sign.enums.CodeMessageEnums;
 import com.webank.webase.sign.exception.BaseException;
 import com.webank.webase.sign.pojo.po.UserInfoPo;
 import com.webank.webase.sign.pojo.vo.ReqEncodeInfoVo;
-import com.webank.webase.sign.pojo.vo.ReqSignMessageHashVo;
+import com.webank.webase.sign.pojo.vo.ReqSignHashVo;
 import com.webank.webase.sign.util.CommonUtils;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
-import org.fisco.bcos.sdk.crypto.CryptoSuite;
-import org.fisco.bcos.sdk.crypto.keypair.CryptoKeyPair;
-import org.fisco.bcos.sdk.crypto.signature.SignatureResult;
-import org.fisco.bcos.sdk.model.CryptoType;
-import org.fisco.bcos.sdk.utils.ByteUtils;
-import org.fisco.bcos.sdk.utils.Hex;
-import org.fisco.bcos.sdk.utils.exceptions.DecoderException;
+import org.fisco.bcos.sdk.v3.crypto.CryptoSuite;
+import org.fisco.bcos.sdk.v3.crypto.keypair.CryptoKeyPair;
+import org.fisco.bcos.sdk.v3.crypto.signature.SignatureResult;
+import org.fisco.bcos.sdk.v3.model.CryptoType;
+import org.fisco.bcos.sdk.v3.utils.Hex;
+import org.fisco.bcos.sdk.v3.utils.exceptions.DecoderException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 /**
@@ -50,6 +50,12 @@ public class SignService {
     ConstantProperties properties;
     @Autowired
     private KeyStoreService keyStoreService;
+    @Autowired
+    @Qualifier(value = "sm")
+    private CryptoSuite smCryptoSuite;
+    @Autowired
+    @Qualifier(value = "ecdsa")
+    private CryptoSuite ecdsaCryptoSuite;
 
     /**
      * add sign.
@@ -61,7 +67,7 @@ public class SignService {
         Instant startTimeDB = Instant.now();
         // check exist
         UserInfoPo userRow = userService.findBySignUserId(signUserId);
-       log.debug("end query db time: {}", Duration.between(startTimeDB, Instant.now()).toMillis());
+        log.debug("end query db time: {}", Duration.between(startTimeDB, Instant.now()).toMillis());
         // check user name not exist.
         if (Objects.isNull(userRow)) {
             log.warn("fail sign, user not exists. signUserId:{}", signUserId);
@@ -71,35 +77,32 @@ public class SignService {
         // signature
         CryptoKeyPair cryptoKeyPair = keyStoreService.getKeyPairByType(userRow.getPrivateKey(), encryptType);
         // make sure hex
-        byte[] encodedData;
         try {
-            encodedData = ByteUtils.hexStringToBytes(req.getEncodedDataStr());
+            Hex.decode(req.getEncodedDataStr());
         } catch (DecoderException e) {
             log.error("hexStringToBytes error: ", e);
             throw new BaseException(CodeMessageEnums.PARAM_ENCODED_DATA_INVALID);
-
         }
         Instant startTime = Instant.now();
         log.info("start sign. startTime:{}", startTime.toEpochMilli());
         // sign message by type
-        SignatureResult signatureResult = signMessageByType(
-                encodedData, cryptoKeyPair, encryptType);
+        SignatureResult signatureResult = signMessageByType(req.getEncodedDataStr(), cryptoKeyPair, encryptType);
         log.info("end sign duration:{}", Duration.between(startTime, Instant.now()).toMillis());
         String signDataStr = CommonUtils.signatureResultToStringByType(signatureResult, encryptType);
         log.info("end sign. signUserId:{}", signUserId);
         return signDataStr;
     }
 
-    public SignatureResult signMessageByType(byte[] message, CryptoKeyPair cryptoKeyPair,
+    public SignatureResult signMessageByType(String messageStr, CryptoKeyPair cryptoKeyPair,
         int encryptType) {
         if (encryptType == CryptoType.SM_TYPE) {
-            byte[] messageHash = new CryptoSuite(CryptoType.SM_TYPE).hash(message);
-            log.debug("userRow.messageHash：{},hex:{}", messageHash, Hex.toHexString(messageHash));
-            return new CryptoSuite(CryptoType.SM_TYPE).sign(Hex.toHexString(messageHash), cryptoKeyPair);
+            String messageHash = smCryptoSuite.hash(messageStr);
+            log.debug("userRow.messageHash：{}", messageHash);
+            return smCryptoSuite.sign(messageHash, cryptoKeyPair);
         } else {
-            byte[] messageHash = new CryptoSuite(CryptoType.ECDSA_TYPE).hash(message);
-            log.debug("userRow.messageHash：{},hex:{}", messageHash, Hex.toHexString(messageHash));
-            return new CryptoSuite(CryptoType.ECDSA_TYPE).sign(Hex.toHexString(messageHash), cryptoKeyPair);
+            String messageHash = ecdsaCryptoSuite.hash(messageStr);
+            log.debug("userRow.messageHash:{}", messageHash);
+            return ecdsaCryptoSuite.sign(messageHash, cryptoKeyPair);
         }
     }
 
@@ -107,7 +110,7 @@ public class SignService {
      * add signHash.
      * @param req parameter
      */
-    public String signMessageHash(ReqSignMessageHashVo req) throws BaseException {
+    public String signMessageHash(ReqSignHashVo req) throws BaseException {
         String signUserId = req.getSignUserId();
         log.info("start sign. signUserId:{}", signUserId);
         Instant startTimeDB = Instant.now();
@@ -126,7 +129,7 @@ public class SignService {
         Instant startTime = Instant.now();
         log.info("start sign. startTime:{}", startTime.toEpochMilli());
         SignatureResult signatureResult = signMessageHashByType(
-                req.getMessageHash(), cryptoKeyPair, encryptType);
+            req.getMessageHash(), cryptoKeyPair, encryptType);
         log.info("end sign duration:{}", Duration.between(startTime, Instant.now()).toMillis());
         String signDataStr = CommonUtils.signatureResultToStringByType(signatureResult, encryptType);
         log.info("end sign. signUserId:{}", signUserId);
@@ -134,11 +137,11 @@ public class SignService {
     }
 
     public SignatureResult signMessageHashByType(String messageHash, CryptoKeyPair cryptoKeyPair,
-                                                 int encryptType) {
+        int encryptType) {
         if (encryptType == CryptoType.SM_TYPE) {
-            return new CryptoSuite(CryptoType.SM_TYPE).sign(messageHash, cryptoKeyPair);
+            return smCryptoSuite.sign(messageHash, cryptoKeyPair);
         } else {
-            return new CryptoSuite(CryptoType.ECDSA_TYPE).sign(messageHash, cryptoKeyPair);
+            return ecdsaCryptoSuite.sign(messageHash, cryptoKeyPair);
         }
     }
 }