+++ src/main/java/com/webank/webase/sign/config/SwaggerConfig.java
@@ -19,7 +19,6 @@ import com.google.common.collect.Lists;
 import java.util.HashSet;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;